requirements.txt
Flask==3.0.0
flask_cors==4.0.1
flask-jwt-extended==4.6.0
PyJWT==2.8.0
cachetools==5.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
requests==2.31.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
mysqlclient==2.2.1
colorama==0.4.6
mysql.connector
stopit==1.1.2
pymysql==1.1.0
pytz
tabulate